requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
jinja2==3.1.2
llama-index==0.9.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pytesseract==0.3.10
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.0.3
python-docx==0.8.11
python-pptx==0.6.21
openpyxl==3.1.2
pdfplumber==0.10.2
pillow==10.0.0
pytest==7.4.0